+++ .env
@@ -1,0 +1,2 @@
+DATA_PATH=/
+CATALOGS=catalog
