+++ Dockerfile
@@ -5,9 +5,10 @@ WORKDIR /app
 COPY requirements.txt /app/requirements.txt
 RUN pip install -r requirements.txt
 COPY . /app
-RUN pip install gunicorn && cd catsHTM && pip install -e .
+RUN cd catsHTM && pip install -e .
 
 EXPOSE 5001
-ENV DATA_PATH="/usr/local/catalogsHTM/"
+ENV DATA_PATH="/data"
+ENV CATALOGS="TMASSxsc,AAVSO_VSX,AKARI,CRTS_per_var,FIRST,NVSS,ROSATfsc,SWIREz"
 WORKDIR /app/scripts
 CMD ["/bin/sh", "entrypoint.sh"]
