+++ requirements.txt
@@ -3,6 +3,7 @@ catsHTM==0.1.28
 Click==7.0
 Flask==1.1.1
 Flask-Cors==3.0.8
+gunicorn
 h5py==2.9.0
 itsdangerous==1.1.0
 Jinja2==2.10.1
